+++ examples/src/labelarg.mli
@@ -1,0 +1,5 @@
+val test : ?fallback:string -> string -> string [@@pure]
+
+val value : 'a option -> default:'a -> 'a
+
+val valueexn : 'a option -> default:'a -> 'a [@@may_raise]